# docs/api/E2-identity/openapi.yaml
openapi: 3.1.0
info:
  title: hris-outsource · E2 Identity, Org & Master Data
  version: 1.0.0
  description: |
    OpenAPI 3.1 contract for **E2 — Identity, Org & Master Data**.

    > **NOTE — Profile change-requests removed 2026-06-14 (EPICS §8 E11):**
    > agent self-edits to phone, emergency contact, and bank account are now
    > **instant** (applied immediately, no approval). The `change-requests`
    > tag, paths, and schemas are deleted. Configurable approval now lives only
    > in E11 (leave/overtime).

    Scope: the people the system is about (employees + linked logins), the
    employment relationship (PKWT/PKWTT agreements + compensation), the client
    company directory, and the operational master data (leave types, attendance
    codes, overtime rules). Position is **free-text** (no master) with a
    typeahead search endpoint over existing values.

    Binding conventions: see `docs/api/CONVENTIONS.md` (sections referenced
    inline as §N). Picker endpoints are co-located with their owning resource
    per §18 — no separate `/lookups` namespace.

    Cross-references: employee-detail tab data (Penempatan/Kehadiran/Cuti/Lembur)
    is **not served** by E2; the E2 employee-detail response carries
    `x-cross-epic-references` mapping each tab to the owning epic's endpoint.

    Personas used in examples: **Sari Hadi (HR Admin)**, **Rudi Wijaya (Shift
    Leader, Plaza Senayan)**, **Budi Santoso (Agent, Parking Attendant)**.
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /agreements:
    get:
      operationId: listAgreements
      tags: [agreements]
      summary: List employment agreements
      description: |
        Filter by employee, status (ACTIVE/SUPERSEDED/CLOSED/EXPIRING), and
        type (PKWT/PKWTT). `EXPIRING` is a server-computed virtual status —
        active PKWT agreements within 30 days of `end_date` (matches the
        notifications cron in CONVENTIONS §16.2).
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      x-design-screens: []  # AUDIT BLOCKER: F2.2 screens entirely missing
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Sort' }
        - in: query
          name: q
          description: Free-text search across employee name, employee id, or agreement number.
          schema: { type: string }
        - in: query
          name: employee_id
          schema: { $ref: '#/components/schemas/EmployeeId' }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/AgreementStatus' }
        - in: query
          name: status__in
          schema: { type: string, example: "ACTIVE,EXPIRING" }
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/AgreementType' }
        - in: query
          name: end_date__lte
          schema: { type: string, format: date }
      responses:
        '200':
          description: Page of agreements.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Agreement' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createAgreement
      tags: [agreements]
      summary: Create a PKWT or PKWTT agreement (EA-1)
      description: |
        Creates an employment agreement. Validations per EA-1/EA-2:
        - `PKWT` requires `start_date` + `end_date`; `PKWTT` requires
          `start_date` only.
        - At most one **active** agreement per employee (EA-2). If one
          exists, either close it first or use `:renew` for the successor
          flow (EA-3).
        - `PKWT` period must be ≤ 5 years (Indonesian labor law) →
          `PKWT_PERIOD_EXCEEDS_MAX` (422).
        - Compensation fields are encrypted at rest (EA-4); see schema
          `CompensationTerms`.
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      x-design-screens: []  # AUDIT BLOCKER
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgreementWriteRequest' }
            examples:
              pkwt_budi:
                summary: PKWT, 1-year, with comp
                value:
                  employee_id: SWP-EMP-1042
                  type: PKWT
                  agreement_no: "PKWT/SWP/2026/0142"
                  start_date: "2026-06-01"
                  end_date: "2027-05-31"
                  compensation:
                    base_salary_idr: 5200000
                    bpjs_terms:
                      kesehatan_employer_pct: 4.0
                      kesehatan_employee_pct: 1.0
                      ketenagakerjaan_employer_pct: 6.24
                      ketenagakerjaan_employee_pct: 3.0
                    tax_profile: PTKP_K0
              pkwtt_rudi:
                summary: PKWTT, indefinite, no end_date
                value:
                  employee_id: SWP-EMP-1043
                  type: PKWTT
                  agreement_no: "PKWTT/SWP/2024/0019"
                  start_date: "2024-01-01"
                  compensation:
                    base_salary_idr: 8500000
                    bpjs_terms:
                      kesehatan_employer_pct: 4.0
                      kesehatan_employee_pct: 1.0
                      ketenagakerjaan_employer_pct: 6.24
                      ketenagakerjaan_employee_pct: 3.0
                    tax_profile: PTKP_K1
      responses:
        '201':
          description: Created.
          headers:
            Location:
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agreement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Active agreement already exists (use `:renew` instead).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                inv2:
                  value:
                    error:
                      code: ACTIVE_AGREEMENT_EXISTS
                      message: "Karyawan sudah memiliki perjanjian aktif. Gunakan endpoint :renew untuk perpanjangan."
                      request_id: req_01J5XKQ9
        '422':
          description: Business-rule violation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                pkwt_too_long:
                  value:
                    error:
                      code: PKWT_PERIOD_EXCEEDS_MAX
                      message: "Periode PKWT melebihi batas 5 tahun yang diizinkan UU Ketenagakerjaan."
                      fields:
                        end_date: "Periode 6 tahun melebihi maksimum 5 tahun."
                      request_id: req_01J5XKR0
                missing_doc:
                  value:
                    error:
                      code: MISSING_REQUIRED_DOCUMENT
                      message: "Dokumen kontrak wajib dilampirkan sebelum perjanjian dapat diaktifkan."
                      request_id: req_01J5XKR1
  /agreements/{agreement_id}:
    parameters:
      - { $ref: '#/components/parameters/AgreementIdPath' }
    get:
      operationId: getAgreement
      tags: [agreements]
      summary: Get agreement detail
      description: |
        Returns the agreement record. **Compensation fields are masked**
        unless the caller has the `view_compensation` permission (EA-4
        role-gated; HR admin & super admin only). Agents (mobile) get a
        slimmer view: type, period, status — compensation hidden (EA-4 §4).
      x-rbac:
        roles: [hr_admin, super_admin, agent]
        scope: self
      responses:
        '200':
          description: Agreement detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agreement' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateAgreement
      tags: [agreements]
      summary: Partial update on an agreement
      description: |
        Patch agreement fields. Compensation changes via this endpoint are
        **effective-dated and historized** as new `CompensationRecord`s (C-3
        decision); the underlying field is never overwritten in place. The
        agreement's "current comp" is the latest record effective today.
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgreementWriteRequest' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agreement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /agreements/{agreement_id}:renew:
    parameters:
      - { $ref: '#/components/parameters/AgreementIdPath' }
    post:
      operationId: renewAgreement
      tags: [agreements]
      summary: Renew a PKWT (generates a linked successor — EA-3)
      description: |
        Creates a **new** agreement with `predecessor_id` set to the current
        one and sets the predecessor's status to `SUPERSEDED`. The successor
        may itself be PKWT (with new end_date) or convert to PKWTT (C-2 —
        permanent conversion).

        Returns the new agreement (with new `SWP-AG-*` id).
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, start_date]
              properties:
                type: { $ref: '#/components/schemas/AgreementType' }
                agreement_no:
                  type: string
                  description: New agreement reference number (e.g. PKWT/SWP/2027/0142).
                start_date: { type: string, format: date }
                end_date:
                  type: string
                  format: date
                  description: Required when type=PKWT.
                compensation: { $ref: '#/components/schemas/CompensationTerms' }
                note:
                  type: string
                  description: Optional human note carried into the audit log.
            examples:
              budi_renew_pkwt:
                value:
                  type: PKWT
                  agreement_no: "PKWT/SWP/2027/0142"
                  start_date: "2027-06-01"
                  end_date: "2028-05-31"
                  compensation:
                    base_salary_idr: 5500000
                    bpjs_terms:
                      kesehatan_employer_pct: 4.0
                      kesehatan_employee_pct: 1.0
                      ketenagakerjaan_employer_pct: 6.24
                      ketenagakerjaan_employee_pct: 3.0
                    tax_profile: PTKP_K0
      responses:
        '201':
          description: Successor created.
          headers:
            Location:
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agreement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Predecessor is not in a renewable state (e.g., already SUPERSEDED).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422':
          description: Business-rule violation (e.g., PKWT > 5 years).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /agreements/{agreement_id}:close:
    parameters:
      - { $ref: '#/components/parameters/AgreementIdPath' }
    post:
      operationId: closeAgreement
      tags: [agreements]
      summary: Close an agreement early (EA-5)
      description: |
        Terminates an agreement with a reason and effective date (EA-5).
        Reason ∈ {RESIGNED, TERMINATED, END_OF_TERM, DECEASED, RETIRED,
        ABSCONDED, OTHER}. Active
        placements (E3) referencing this agreement are flagged for review
        (cascade — implicit side effect; not a separate API call).
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, effective_date]
              properties:
                reason:
                  type: string
                  enum: [RESIGNED, TERMINATED, END_OF_TERM, DECEASED, RETIRED, ABSCONDED, OTHER]
                effective_date:
                  type: string
                  format: date
                note:
                  type: string
                  maxLength: 500
            examples:
              budi_resigned:
                value:
                  reason: RESIGNED
                  effective_date: "2026-08-15"
                  note: "Pengunduran diri atas permintaan sendiri (lampirkan surat di attachments)."
      responses:
        '200':
          description: Closed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agreement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already closed/superseded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-jwt
      description: |
        `Authorization: Bearer <token>` per CONVENTIONS §3. Token type
        (opaque vs JWT) is deferred to E1; consumers MUST treat as opaque.
  parameters:
    Limit:
      in: query
      name: limit
      description: Page size (default 50, max 200) per §8.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      in: query
      name: cursor
      description: Opaque cursor from a previous page's `next_cursor` per §8.
      schema: { type: string }
    Sort:
      in: query
      name: sort
      description: |
        Comma-separated `field:direction` per §9 (e.g.,
        `?sort=created_at:desc,full_name:asc`). Cursor encodes sort — mismatch
        returns `CURSOR_MISMATCH` (400).
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: UUID v4. Same key with different body → `IDEMPOTENCY_KEY_REUSED` (409). Per §13.
      schema: { type: string, format: uuid }
...
    AgreementIdPath:
      in: path
      name: agreement_id
      required: true
      schema: { $ref: '#/components/schemas/AgreementId' }
...
  responses:
    BadRequest:
      description: Malformed request or invalid input (CONVENTIONS §11).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            generic:
              value:
                error:
                  code: INVALID_REQUEST
                  message: "Permintaan tidak valid."
                  request_id: req_01J5XK0A
    Unauthenticated:
      description: Missing or expired bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            expired:
              value:
                error:
                  code: UNAUTHENTICATED
                  message: "Sesi berakhir. Silakan masuk kembali."
                  request_id: req_01J5XK0B
    Forbidden:
      description: Authenticated but role-not-permitted (CONVENTIONS §3, §17).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            no_permission:
              value:
                error:
                  code: FORBIDDEN
                  message: "Anda tidak memiliki izin untuk tindakan ini."
                  request_id: req_01J5XK0C
    NotFound:
      description: Resource not found (or hidden from caller's scope).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            generic:
              value:
                error:
                  code: NOT_FOUND
                  message: "Data tidak ditemukan."
                  request_id: req_01J5XK0D
    UnprocessableEntity:
      description: Business-rule violation (semantically invalid).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            generic:
              value:
                error:
                  code: RULE_VIOLATION
                  message: "Aturan bisnis tidak terpenuhi."
                  request_id: req_01J5XK0E
  schemas:
...
    PageEnvelope:
      type: object
      description: Standard pagination envelope (CONVENTIONS §8).
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: [string, "null"]
          description: Opaque cursor for the next page; `null` when no more pages.
        has_more:
          type: boolean
      required: [data, has_more]
    ErrorEnvelope:
      type: object
      description: Standard error envelope (CONVENTIONS §11).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: "QUOTA_EXCEEDED" }
            message: { type: string, example: "Permintaan melebihi kuota tersisa." }
            fields:
              type: object
              additionalProperties: { type: string }
              description: Only on 400/422 — maps field name → field-level error.
            request_id: { type: string, example: "req_01J5XK..." }
...
    EmployeeId:
      type: string
      pattern: '^SWP-EMP-\d+$'
      example: SWP-EMP-1042
    AgreementId:
      type: string
      pattern: '^SWP-AG-\d+$'
      example: SWP-AG-7711
...
    AgreementType:
      type: string
      enum: [PKWT, PKWTT]
      description: |
        Indonesian labor-law employment type. `PKWT` = perjanjian kerja
        waktu tertentu (fixed-term, with end_date). `PKWTT` = perjanjian
        kerja waktu tidak tertentu (indefinite, no end_date).
    AgreementStatus:
      type: string
      enum: [ACTIVE, SUPERSEDED, CLOSED, EXPIRING]
      description: |
        `EXPIRING` is a server-computed virtual status — active PKWT within
        30 days of `end_date`. Persisted status is one of ACTIVE/SUPERSEDED/CLOSED.
    AgreementCloseReason:
      type: string
      enum: [RESIGNED, TERMINATED, END_OF_TERM, DECEASED, RETIRED, ABSCONDED, OTHER]
    TaxProfile:
      type: string
      description: |
        Indonesian PTKP code. Examples: `PTKP_TK0` (single, no dependents),
        `PTKP_K0`, `PTKP_K1` (married, N dependents).
      example: PTKP_K0
    BpjsTerms:
      type: object
      description: BPJS percentages (statutory + agreement-specific overrides).
      properties:
        kesehatan_employer_pct: { type: number, format: float, example: 4.0 }
        kesehatan_employee_pct: { type: number, format: float, example: 1.0 }
        ketenagakerjaan_employer_pct: { type: number, format: float, example: 6.24 }
        ketenagakerjaan_employee_pct: { type: number, format: float, example: 3.0 }
    CompensationTerms:
      type: object
      description: |
        Current compensation. Encrypted at rest (EA-4); response **masks**
        values to `***` for callers without the `view_compensation`
        permission (HR admin & super admin only; agents on mobile never see
        these — they get the agreement summary only).
      properties:
        base_salary_idr:
          type: number
          format: float
          description: Base monthly salary in IDR. Encrypted at rest.
          example: 5200000
        bpjs_terms: { $ref: '#/components/schemas/BpjsTerms' }
        tax_profile: { $ref: '#/components/schemas/TaxProfile' }
        effective_date:
          type: string
          format: date
          description: |
            Effective date of these terms (`CompensationRecord.effective_date`
            — F2.2 C-3). Defaults to agreement start_date on create.
    Agreement:
      type: object
      required: [id, employee_id, type, status, start_date]
      properties:
        id: { $ref: '#/components/schemas/AgreementId' }
        employee_id: { $ref: '#/components/schemas/EmployeeId' }
        employee_name:
          type: string
          nullable: true
          description: >-
            Full name of the agreement's employee (joined from the employee
            record). Populated on the list projection for display; may be
            absent if the employee record is unavailable.
          example: "Budi Santoso"
        type: { $ref: '#/components/schemas/AgreementType' }
        agreement_no:
          type: string
          example: "PKWT/SWP/2026/0142"
        start_date: { type: string, format: date }
        end_date:
          type: [string, "null"]
          format: date
          description: Required when type=PKWT; null for PKWTT.
        status: { $ref: '#/components/schemas/AgreementStatus' }
        predecessor_id:
          allOf: [{ $ref: '#/components/schemas/AgreementId' }]
          nullable: true
          description: Set on the successor created via `:renew` (EA-3).
        successor_id:
          allOf: [{ $ref: '#/components/schemas/AgreementId' }]
          nullable: true
        closed_reason: { $ref: '#/components/schemas/AgreementCloseReason' }
        closed_at:
          type: [string, "null"]
          format: date-time
        compensation:
          allOf: [{ $ref: '#/components/schemas/CompensationTerms' }]
          description: |
            Current comp. Hidden / masked for unauthorized roles (EA-4). On
            mobile (agent role) this field is omitted entirely.
        annual_leave_entitlement_days:
          type: integer
          minimum: 0
          nullable: true
          description: >-
            Annual leave entitlement in days (hak cuti tahunan) — term of the
            employment agreement; source for E6 leave quota.
          example: 12
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
    AgreementWriteRequest:
      type: object
      description: Body for create/patch on agreements.
      properties:
        employee_id: { $ref: '#/components/schemas/EmployeeId' }
        type: { $ref: '#/components/schemas/AgreementType' }
        agreement_no: { type: string }
        start_date: { type: string, format: date }
        end_date:
          type: string
          format: date
          description: Required for PKWT; must be omitted for PKWTT.
        compensation: { $ref: '#/components/schemas/CompensationTerms' }
        annual_leave_entitlement_days:
          type: integer
          minimum: 0
          nullable: true
          description: >-
            Annual leave entitlement in days (hak cuti tahunan) — term of the
            employment agreement; source for E6 leave quota.

    # ---- ClientCompany --------------------------------------------------
...